controller

+++ src/main/java/com/gardenmap/gardenmap/model/Product.java
@@ -5,10 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 
 @Entity
@@ -17,7 +14,7 @@ import javax.persistence.Table;
 public class Product {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private int price;
